+++ src/System.Net.Http/ref/System.Net.Http.csproj
@@ -10,6 +10,7 @@
   <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'uap-Release|AnyCPU'" />
   <ItemGroup>
     <Compile Include="System.Net.Http.cs" />
+    <Compile Include="System.Net.Http.SocketsHttpHandler.cs" Condition="'$(TargetGroup)' != 'uap'" />
   </ItemGroup>
   <ItemGroup Condition="'$(TargetGroup)' == 'netcoreapp' or '$(TargetGroup)' == 'uap'">
     <ProjectReference Include="..\..\System.Runtime\ref\System.Runtime.csproj" />
