+++ src/System.Net.Http/src/System/Net/Http/HttpClientHandler.Core.cs
@@ -62,12 +62,5 @@ namespace System.Net.Http
                 throw new InvalidOperationException(SR.net_http_operation_started);
             }
         }
-
-        private void ThrowForModifiedManagedSslOptionsIfStarted()
-        {
-            // Hack to trigger an InvalidOperationException if a property that's stored on
-            // SslOptions is changed, since SslOptions itself does not do any such checks.
-            _socketsHttpHandler.SslOptions = _socketsHttpHandler.SslOptions;
-        }
     }
 }
